Move defaultJourneyModes from hooks to utils to unravel dependencies..
