--- src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/Dashboard_TripSearchScreen/are-default-filters-selected.ts
@@ -1,8 +1,0 @@
-import {TransportModeFilterOptionWithSelectionType} from '@atb/travel-search-filters';
-
-export const areDefaultFiltersSelected = (
-  transportModes?: TransportModeFilterOptionWithSelectionType[],
-): boolean => {
-  if (!transportModes) return false;
-  return !transportModes.some((tm) => tm.selectedAsDefault !== tm.selected);
-};
